chore: seperated code from column into its own component actions
