+++ src/modules/admin/ui/components/actions.tsx
@@ -1,0 +1,103 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+
+import { Ellipsis, Edit, Trash2 } from "lucide-react";
+
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useTRPC } from "@/trpc/client";
+
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
+
+interface ActionsProps {
+  userId: string;
+  filters: {
+    search: string;
+    page: number;
+    memberId: string;
+  };
+  confirmDelete: () => Promise<boolean>;
+}
+
+export const Actions = ({ userId, filters, confirmDelete }: ActionsProps) => {
+  const trpc = useTRPC();
+  const queryClient = useQueryClient();
+
+  // const updateUser = useMutation(
+  //   trpc.profile.adminEdit.mutationOptions({
+  //     onSuccess: async () => {
+  //       await queryClient.invalidateQueries(
+  //         trpc.members.getMany.queryOptions({ ...filters }),
+  //       );
+  //     },
+  //     onError: (error) => console.error(error.message),
+  //   }),
+  // );
+
+  const deleteUser = useMutation(
+    trpc.profile.adminDelete.mutationOptions({
+      onSuccess: async () => {
+        await queryClient.invalidateQueries(
+          trpc.members.getMany.queryOptions({ ...filters }),
+        );
+      },
+      onError: (error) => console.error(error.message),
+    }),
+  );
+
+  const onDelete = async () => {
+    const Ok = await confirmDelete();
+
+    if (!Ok) {
+      return;
+    }
+
+    deleteUser.mutate({ userId });
+  };
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button
+          variant="ghost"
+          size="sm"
+          className="h-8 w-8 p-0 hover:bg-zinc-800/50"
+        >
+          <span className="sr-only">Open menu</span>
+          <Ellipsis className="h-4 w-4" />
+        </Button>
+      </DropdownMenuTrigger>
+
+      <DropdownMenuContent
+        align="start"
+        className="w-48 border-zinc-800 bg-zinc-900/95 backdrop-blur-sm"
+      >
+        <DropdownMenuGroup>
+          <DropdownMenuItem className="group cursor-pointer text-zinc-300 focus:bg-zinc-800">
+            <Edit className="focus:group-text-zinc-300 mr-2 h-4 w-4" />
+            <span className="focus:group-text-zinc-300">Edit Member</span>
+          </DropdownMenuItem>
+        </DropdownMenuGroup>
+
+        <DropdownMenuSeparator className="bg-zinc-800" />
+
+        <DropdownMenuGroup>
+          <DropdownMenuItem
+            className="group cursor-pointer text-red-400 focus:bg-red-500/20 focus:text-red-600"
+            onClick={onDelete}
+          >
+            <Trash2 className="focus:group-text-red-600 mr-2 h-4 w-4" />
+            <span className="focus:group-text-red-600">Delete Member</span>
+          </DropdownMenuItem>
+        </DropdownMenuGroup>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+};
